New input processing
Rather than have the input processing spread out everywhere, I added a new function to read any input message and process it into flags and data that make the logic look a lot nicer. It might also reduce the amount of processing we need to do on any given input because its all "centralized" and things are guaranteed to not repeat themselves.