Lots of updates
Redid all inputs and outputs to play with numbers nicer. This involved implementing the parseInt() function.
Added commas to separate parts of the message due to a lack of indexing desired characters.
Fixed offset so the nodes should be able to match the global cycle.
Base node can now read messages.